人面不知何处去，桃花依旧笑春风
战场的人物层改用指针，不用编号，ai计算的时候会假象一个移动过的临时人物，修改几个=写成==的地方

+++ src/Cloud.h
@@ -17,7 +17,7 @@ public:
     };
 
     Point position_;
-    float speed_;
+    int speed_;
 
     const int max_X = 17280;
     const int max_Y = 8640;
